requirements.txt
Django==3.2.10
django-crontab==0.7.1
django-environ==0.8.1
djangorestframework==3.13.1
